# frontend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 慧眼医疗云平台 V2.0 API
  description: |
    糖尿病视网膜病变 AI 辅助筛查 + 眼科医学影像培训考核

    **后端框架推荐**：FastAPI / Spring Boot / NestJS
    **协议**：REST + JSON · UTF-8
    **认证**：`Authorization: Bearer <token>`

    ## 统一响应结构

    所有接口（除文件下载、图片流外）统一返回：

    ```json
    {
      "code": 0,
      "msg": "success",
      "data": {}
    }
    ```

    - `code === 0` 或 `200` 表示成功
    - `code === 401` 或 `40101` 表示会话失效，前端自动跳登录
    - 其它非零 `code` 表示业务失败，前端使用 `msg` 弹错误提示

    ## 错误状态码

    | HTTP | 含义 |
    |---|---|
    | 400 | 请求参数错误 |
    | 401 | 未登录 / 会话失效 |
    | 403 | 无权限 |
    | 404 | 资源不存在 |
    | 413 | 上传文件过大 |
    | 500 | 服务器错误 |
    | 502/503/504 | 网关/服务不可用 |
  version: 2.0.0
  contact:
    name: 慧眼医疗 · 前端工程组
  license:
    name: Proprietary
servers:
  - url: http://127.0.0.1:8000
    description: 本地开发
  - url: https://huiyan.example.com/api
    description: 生产（Nginx 反向代理）
...
security:
  - bearerAuth: []
paths:
...
  /screening/upload:
    post:
      tags: [Screening · 体检筛查]
      summary: 单张眼底图上传
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 眼底图（JPG/PNG/BMP，≤ 20MB）
                patientId: { type: string, example: P2026001 }
                patientName: { type: string, example: 王建国 }
                gender: { type: string, enum: ['男', '女'] }
                age: { type: integer, example: 64 }
                eye: { type: string, enum: [OD, OS, OU] }
                hospital: { type: string }
                doctor: { type: string }
                remark: { type: string }
      responses:
        '200':
          description: 上传成功，返回任务 ID
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - properties:
                      data:
                        $ref: '#/components/schemas/UploadFundusResult'
  /screening/upload/batch:
    post:
      tags: [Screening · 体检筛查]
      summary: 批量上传眼底图（支持文件夹）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                hospital: { type: string }
                doctor: { type: string }
      responses:
        '200':
          description: 批量上传结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - properties:
                      data:
                        $ref: '#/components/schemas/BatchUploadResult'
  /screening/tasks:
    get:
      tags: [Screening · 体检筛查]
      summary: 获取筛查任务分页列表
      parameters:
        - { name: keyword, in: query, schema: { type: string }, description: 关键词（患者ID/姓名/文件名） }
        - { name: risk, in: query, schema: { $ref: '#/components/schemas/RiskLevel' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ScreeningStatus' } }
        - { name: startTime, in: query, schema: { type: string, example: '2026-05-01 00:00:00' } }
        - { name: endTime, in: query, schema: { type: string, example: '2026-05-21 23:59:59' } }
        - { name: hospital, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 50 } }
        - { name: sortBy, in: query, schema: { type: string, enum: [risk, createdAt, confidence] } }
        - { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: 任务分页
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageResult'
                          - properties:
                              list:
                                type: array
                                items: { $ref: '#/components/schemas/ScreeningTask' }
  /screening/tasks/{taskId}:
    get:
      tags: [Screening · 体检筛查]
      summary: 获取单个筛查任务详情
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - properties:
                      data:
                        $ref: '#/components/schemas/ScreeningTask'
    delete:
      tags: [Screening · 体检筛查]
      summary: 移除筛查任务
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
  /screening/tasks/reanalyze:
    post:
      tags: [Screening · 体检筛查]
      summary: 重新分析任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskId]
              properties:
                taskId: { type: string }
                force: { type: boolean, default: false, description: 忽略缓存 }
      responses:
        '200':
          description: 已重新加入分析队列
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - properties:
                      data:
                        $ref: '#/components/schemas/ScreeningTask'
...
  /screening/reports/{taskId}:
    get:
      tags: [Screening · 体检筛查]
      summary: 获取筛查报告详情
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 报告详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - properties:
                      data:
                        $ref: '#/components/schemas/ScreeningReport'
  /screening/reports/{taskId}/pdf:
    get:
      tags: [Screening · 体检筛查]
      summary: 导出单份筛查报告 PDF
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: PDF 文件流
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /screening/reports/summary/pdf:
    get:
      tags: [Screening · 体检筛查]
      summary: 导出汇总报告 PDF（按筛选条件）
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: risk, in: query, schema: { $ref: '#/components/schemas/RiskLevel' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ScreeningStatus' } }
      responses:
        '200':
          description: PDF 文件流
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /screening/tasks/export:
    get:
      tags: [Screening · 体检筛查]
      summary: 导出筛查任务 Excel
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: risk, in: query, schema: { $ref: '#/components/schemas/RiskLevel' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ScreeningStatus' } }
      responses:
        '200':
          description: Excel 文件流
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /screening/refer:
    post:
      tags: [Screening · 体检筛查]
      summary: 高危病例一键转诊
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskId, targetHospital]
              properties:
                taskId: { type: string }
                targetHospital: { type: string, example: 北京同仁医院眼科 }
                note: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

  # ============================================================
  #                  Training · 医学培训
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 在 `Authorization` 头中以 `Bearer <token>` 形式提交
  responses:
    EmptySuccess:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResult'
  schemas:
...
    ApiResult:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          example: 0
          description: 业务码（0/200=成功，401=未登录，其它=失败）
        msg:
          type: string
          example: success
        data:
          description: 业务数据，按各接口具体字段
          nullable: true
    PageResult:
      type: object
      required: [total, page, pageSize, list]
      properties:
        total: { type: integer, example: 128 }
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 20 }
        list:
          type: array
          items: { type: object }

    # ===================== 枚举 =====================
    RiskLevel:
      type: string
      enum: [red, yellow, green]
      description: 风险等级 — red 高危转诊 / yellow 中危随访 / green 正常
    ScreeningStatus:
      type: string
      enum: [queued, analyzing, done, failed]
      description: 任务状态
    EyeSide:
      type: string
      enum: [OD, OS, OU]
      description: OD=右眼 / OS=左眼 / OU=双眼
    Gender:
      type: string
      enum: ['男', '女']
...
    UploadFundusResult:
      type: object
      required: [taskId, fileUrl, fileName, fileSize, queued]
      properties:
        taskId: { type: string, example: T2026001 }
        fileUrl: { type: string, format: uri }
        fileName: { type: string, example: P2026001_OD.jpg }
        fileSize: { type: integer, example: 2154321, description: 字节数 }
        queued: { type: boolean }
    BatchUploadResult:
      type: object
      required: [total, success, failed, items]
      properties:
        total: { type: integer }
        success: { type: integer }
        failed: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/UploadFundusResult' }
    ScreeningTask:
      type: object
      required: [id, patientId, patientName, eye, age, gender, status, dr, confidence, createdAt, fileName, hospital, doctor, remark]
      properties:
        id: { type: string, example: T2026001 }
        patientId: { type: string, example: P2026001 }
        patientName: { type: string, example: 王建国 }
        eye: { $ref: '#/components/schemas/EyeSide' }
        age: { type: integer, example: 64 }
        gender: { $ref: '#/components/schemas/Gender' }
        status: { $ref: '#/components/schemas/ScreeningStatus' }
        risk:
          allOf: [{ $ref: '#/components/schemas/RiskLevel' }]
          nullable: true
        dr: { type: string, example: '4 级 PDR（增殖性）' }
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.974
        createdAt: { type: string, example: '2026-05-21 08:42:15' }
        fileName: { type: string, example: P2026001_OU.jpg }
        fileUrl: { type: string, format: uri }
        thumbUrl: { type: string, format: uri }
        hospital: { type: string }
        doctor: { type: string }
        remark: { type: string }
...
    ScreeningReport:
      type: object
      required: [taskId, patientId, patientName, gender, age, eye, hospital, doctor, examTime, reportTime, reportNo, risk, dr, confidence, conclusion, suggestion, lesions, imageUrls]
      properties:
        taskId: { type: string }
        patientId: { type: string }
        patientName: { type: string }
        gender: { $ref: '#/components/schemas/Gender' }
        age: { type: integer }
        eye: { $ref: '#/components/schemas/EyeSide' }
        hospital: { type: string }
        doctor: { type: string }
        examTime: { type: string }
        reportTime: { type: string }
        reportNo: { type: string, example: HY-2026-05-21-0001 }
        risk: { $ref: '#/components/schemas/RiskLevel' }
        dr: { type: string }
        confidence: { type: number, format: float }
        conclusion: { type: string }
        suggestion: { type: string }
        lesions:
          type: array
          items:
            type: object
            properties:
              type: { type: string }
              count: { type: integer }
              location: { type: string }
        imageUrls:
          type: object
          properties:
            origin: { type: string, format: uri }
            heatmap: { type: string, format: uri }
        reviewer: { type: string }
        reviewedAt: { type: string }

    # ===================== 医学培训 =====================
...